moved testDbConnection to index.ts and combined the user_signup.ts file and user_login.ts file into one file named auth.ts file removing too much modularity
